Разделил код на модули: main, util, data
